refactor: assign negative item identifier when an item is created
When a `RegisterItem` is first created (e.g. as part of an addition proposal),
the system will now assign a temporary negative item identifier. This negative
identifier will be overwritten by a final positive identifier once the item
is accepted into the register.

This change does not affect existing items. It allows, however, for a more
consistent way to handle item identifier assignments when continuous
identifiers are desired for the registered items.